Final V1.7

+++ source/external_func/ft_lstsize.c
@@ -6,7 +6,7 @@
 /*   By: ysbai-jo <ysbai-jo@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 18:34:33 by ysbai-jo          #+#    #+#             */
-/*   Updated: 2024/04/27 12:22:39 by ysbai-jo         ###   ########.fr       */
+/*   Updated: 2024/05/01 10:27:43 by ysbai-jo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 unsigned int	ft_lstsize(t_stack *lst)
 {
-	unsigned int	s;
+	size_t	s;
 
 	s = 0;
 	while (lst)
